# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Your API Title",
    "version": "1.0.0",
    "description": "Your API description"
  },
  "paths": {
    "/customers": {
      "get": {
        "summary": "Retrieve all customers",
        "tags": [
          "Customers"
        ],
        "responses": {
          "200": {
            "description": "A list of customers"
          }
        }
      },
      "post": {
        "summary": "Create a new customer",
        "tags": [
          "Customers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the customer"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address of the customer"
                  }
                },
                "example": {
                  "name": "John Doe",
                  "email": "john.doe@example.com"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created customer object"
          }
        }
      }
    },
    "/customers/{id}": {
      "get": {
        "summary": "Retrieve a single customer by ID",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the customer to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "A customer object"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      },
      "put": {
        "summary": "Update a customer by ID",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the customer to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Updated name of the customer"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Updated email address of the customer"
                  }
                },
                "example": {
                  "name": "John Doe (Updated)",
                  "email": "john.doe@example.com"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated customer object"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a customer by ID",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the customer to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Customer deleted successfully"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      }
    },
    "/orders": {
      "get": {
        "summary": "Retrieve all orders",
        "tags": [
          "Orders"
        ],
        "responses": {
          "200": {
            "description": "A list of orders"
          }
        }
      },
      "post": {
        "summary": "Create a new order",
        "tags": [
          "Orders"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customerId": {
                    "type": "integer",
                    "description": "ID of the customer placing the order"
                  },
                  "amount": {
                    "type": "number",
                    "description": "Total amount of the order"
                  }
                },
                "example": {
                  "customerId": 1,
                  "amount": 100.5
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created order object"
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Retrieve a single order by ID",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the order to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "An order object"
          },
          "404": {
            "description": "Order not found"
          }
        }
      },
      "put": {
        "summary": "Update an order by ID",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the order to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customerId": {
                    "type": "integer",
                    "description": "Updated ID of the customer placing the order"
                  },
                  "amount": {
                    "type": "number",
                    "description": "Updated total amount of the order"
                  }
                },
                "example": {
                  "customerId": "1 (Updated)",
                  "amount": 120.75
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated order object"
          },
          "404": {
            "description": "Order not found"
          }
        }
      },
      "delete": {
        "summary": "Delete an order by ID",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the order to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Order deleted successfully"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
    "/product-categories": {
      "get": {
        "summary": "Retrieve all product categories",
        "tags": [
          "Product Categories"
        ],
        "responses": {
          "200": {
            "description": "A list of product categories"
          }
        }
      },
      "post": {
        "summary": "Create a new product category",
        "tags": [
          "Product Categories"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the product category"
                  }
                },
                "example": {
                  "name": "Electronics"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created product category object"
          }
        }
      }
    },
    "/product-categories/{id}": {
      "get": {
        "summary": "Retrieve a single product category by ID",
        "tags": [
          "Product Categories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the product category to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "A product category object"
          },
          "404": {
            "description": "Product category not found"
          }
        }
      },
      "put": {
        "summary": "Update a product category by ID",
        "tags": [
          "Product Categories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the product category to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Updated name of the product category"
                  }
                },
                "example": {
                  "name": "Electronics (Updated)"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated product category object"
          },
          "404": {
            "description": "Product category not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a product category by ID",
        "tags": [
          "Product Categories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the product category to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Product category deleted successfully"
          },
          "404": {
            "description": "Product category not found"
          }
        }
      }
    },
...
  },
...
}